Add an ATON constraint to the time step calculation.
In Iliev Test 5, the gas properties are initially constant so the hydro
time step is really big (200 Myr). If we just use that time step for
ATON, then the coupling doesn't work properly because it has too many
substeps before the next hydro step.

This is only a problem in the Iliev tests because of the gas initial
conditions. It's not a problem for cosmo runs. So I just created a
namelist parameter that can be set manually to limit the time step if
needed.

Tested:
Ran test 5.


git-svn-id: https://svn.physik.uzh.ch/repos/itp/ramses@140 fa3a7ada-b1ec-11df-bea4-b5d77251c905