OpenMP: hydro on unigrid (#119)
This PR adds full openMP support for hydro simulations on a uniform
grid, as benchmarked by the sedov3d benchmark.
OpenMP is added to:
* main loop of godunov solver
* take into account atomic updates in the case of AMR in godfine1
* main loops of set_uold and set_unew
* main loop of courant_fine
* main loop of upload_fine

For a 1024^3 sedov benchmark, a speedup can be observed from 64 nodes
when using 8 threads (as measured on MeluXina).
The main advantage for using openMP in unigrid hydro simulation comes
from the major reduction in memory requirements to store the computation
domain, due to there being less boundary grids. For example, on 32 nodes
with 4 threads the memory required is only a third of what is was with
mpi-only.

This work was done is the context of the SPACE CoE and based in part on
the yOMP version of Ramses developed by San Han.