create calc_uslope subroutines for each slope_type
I've moved the if-check inwards, so that the outer loop i,j,k is now only done once. Then I created vectorized subroutines for each slope_type, so that uslope is now effectively a wrapper routine to select the requested slope type.
By keeping the dimensions of input and output arrays the same as before, the compiler properly inlines and vectorizes. And so the execution time is unaffected.

This structure will allow to move the call to uslope inside trace, bypassing a slow store-load operation.