Refactor and optimize uslope (#154)
I've gather the formulas for the different slope limiters into a module
called slope_types. For each slope type, there is a pure function which
the compiler should inline thanks to the additional compiler flags in
the Makefile.
This structure reduces code duplication and makes it easier to optimize
the existing slope limiters.

I've rewritten minmod/average for avoid the division and casting from
int to float in the inner loop. This also makes the calculation of dlft
and drgt the same as for the other regular slope types.

The slope module also contains a subroutine for each slope_type which is
doing the calculation for nvector elements. Most of these call their
corresponding slope_type function for each dimension.
In these subroutines, the center value is now prefetched into a scalar,
and reused for the calculation of dlft and drgt, reducing the number of
load operations. This gives a small performance gain.

The main subroutine uslope is now effectively a wrapper to call the
requested calc_uslope subroutine.
I've rearranged the parts for different NDIM to have less duplication.
For HYDRO slope_type 1, the structure is a bit strange due to their
being a different formulation for 3D for unclear reasons. Since I didn't
want to change the functionality, I've kept both.
 
For MHD, the calculation of the B-field slopes are now gathered at the
end of the routine, to make it easier to merge hydro and mhd files in
the future.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Noé Brucy <nbrucy@users.noreply.github.com>
Co-authored-by: Noe Brucy <noe.brucy@ens-lyon.fr>
Co-authored-by: Romain Teyssier <romain.teyssier@gmail.com>