optimizations for hydro solver on uniform grid (#159)
* Add the constants defining the loop bounds for the hydro stencil to
the amr_constants module
* Move the gathering of the stencil from godfine1 to its own routine
"gather_stencil_amr"
* Add a version of gather_stencil for uniform grid runs. 
 
For unigrid simulations, this circumvents the non-vectorizable indirect
access caused by splitting up the cells into those that need
interpolation and those that don't. We're still working on a solution
for AMR runs.

This changes results in a 10% speedup for unigrid hydro simulations.

![check_refactor](https://github.com/user-attachments/assets/30c9d7a8-89e9-4d84-b07a-cd43a73ce367)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Romain Teyssier <romain.teyssier@gmail.com>