Refactor and document trace (#170)
Full refactoring of the trace routines for the umuscl scheme in the
godunov solver (hydro solver only).
* Combine trace1d, trace2d and trace3d into a single trace routine,
removing code duplication.
* Move loop over nvar inside the vectorization loop, allowing for
generalized code. These loops are unrolled as to behave similarly to the
scalar instructions from before.
* Add loops over ndim, allowing to remove preprocessor directives. These
loops are also unrolled.
* SIMD compiler directive added to ensure the vectorization is done on
the correct loop. Despite this, the vectorization rate of the trace
routine is reduced. Attempts at restoring the vectorization resulted in
performance loss.
* Rewrite the calculations to reduce the number of math operations:
    * use only dtdx since dx=dy=dz.
    * precalculate /r once, then reuse.
* Add comments describing each term in the calculation.
* Add documentation at the top of the routine.

In the end, the performance is the same as the old version for the sedov
benchmark, when measure on Meluxina with GNU.

WARNING: this PR is touching a performance heavy part of the hydro
solver. If you experience any problems with compilation or performance
related to this PR, please contact me.

---------

Co-authored-by: Tine Colman <tine.colman@ens-lyon.fr>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Romain Teyssier <romain.teyssier@gmail.com>